requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
openai
deepgram-sdk==3.*
#git+https://github.com/m-bain/whisperx.git